pkg/users: fix tests
--- sturdy ---
user_id: cc33c18b-27e5-470d-8589-f9391ca37590
view_id: 2c064d23-2410-4e86-939f-91c56dae2fae
workspace_id: 669ca7a1-5988-4706-b756-cfffde162e45
