+++ api/pkg/users/db/db.go
@@ -10,17 +10,6 @@ import (
 	"github.com/lib/pq"
 )
 
-type Repository interface {
-	Create(newUser *users.User) error
-	Get(id string) (*users.User, error)
-	GetByIDs(ctx context.Context, ids ...string) ([]*users.User, error)
-	GetByEmail(email string) (*users.User, error)
-	Update(*users.User) error
-	UpdatePassword(u *users.User) error
-	Count(context.Context) (uint64, error)
-	List(ctx context.Context, limit uint64) ([]*users.User, error)
-}
-
 type repo struct {
 	db *sqlx.DB
 }
